+++ tests/integration/ante/test_evm_unit_10_gas_wanted.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/cosmos/evm/ante/evm"
-	testconstants "github.com/cosmos/evm/testutil/constants"
+	testconfig "github.com/cosmos/evm/testutil/config"
 	"github.com/cosmos/evm/testutil/integration/evm/factory"
 	"github.com/cosmos/evm/testutil/integration/evm/grpc"
 	"github.com/cosmos/evm/testutil/integration/evm/network"
@@ -20,12 +20,16 @@ import (
 
 func (s *EvmUnitAnteTestSuite) TestCheckGasWanted() {
 	keyring := testkeyring.New(1)
-	unitNetwork := network.NewUnitTestNetwork(
-		s.create,
-		network.WithChainID(testconstants.ChainID{
+	chainConfig := testconfig.ChainConfig{
+		ChainInfo: testconfig.ChainInfo{
 			ChainID:    s.ChainID,
 			EVMChainID: s.EvmChainID,
-		}),
+		},
+		CoinInfo: testconfig.DefaultChainConfig.CoinInfo,
+	}
+	unitNetwork := network.NewUnitTestNetwork(
+		s.create,
+		network.WithChainConfig(chainConfig),
 		network.WithPreFundedAccounts(keyring.GetAllAccAddrs()...),
 	)
 	grpcHandler := grpc.NewIntegrationHandler(unitNetwork)