+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - [\#568](https://github.com/cosmos/evm/pull/568) Avoid unnecessary block notifications when the event bus is already set up.
 - [\#511](https://github.com/cosmos/evm/pull/511) Minor code cleanup for `AddPrecompileFn`.
 - [\#544](https://github.com/cosmos/evm/pull/544) Parse logs from the txResult.Data and avoid emitting EVM events to cosmos-sdk events.
+- [\#436](https://github.com/cosmos/evm/pull/436) Refactor and consolidate chain config helpers and constants for test suites
+>>>>>>> main
 
 ### FEATURES